Fix dragging apparent delay issue
Dragging (especially on mobile) with our custom dragging logic felt
slow, but it's actually a bug. Upon render we don't take the occasion to
update the values; we still let the raf (which might happen a few ms
later) do it.

This fixes it by simply updating the value in `willUpdate`. Might have
some potential optimizations here when we bring back the animation loop
and take the occasion to run a tick or something.