Try new dependent spring API on demo1 (chat heads)
Temporarily revert to not using defaultValue. Will put it back later.
We also use context + cursor (`<Child to={1}`) to pass things down and
refer to things, to keep the perf opt of not re-rendering. Might switch
back to using child fn + passing down configs. The problem with this
approach is the man-in-the-middle attack where we subsequently add
another `Springs` wrapping the desired `Child`. We accidentally
overshadow the original context.

Ping me if you're still reading this because I can't imagine anyone
reading this right now =). Also, if you can solve this context problem
I'll buy you a beer.